requirements.txt
pillow == 5.2.0
twine >= 1.8.0
setuptools >= 38.6.0
wheel
tox
pytest
mypy
black
